refactor: HandlerAdapter 패키지 변경
